refactor(core): SearchEngine の構築処理を search/build.rs へ分割 (#598)
検索ホットパスとインデックス構築時のみ動く責務が search.rs に同居していた。
構築処理を private submodule `search/build.rs` へ移設し、製品挙動は変えない。

- 移設: `Wave1Strings` / `compute_wave1` / `compute_wave2` /
  `compute_kana_char_masks` と、全コンストラクタ（`assemble` / `new` /
  `new_with_migemo` / `new_with_cached_masks`）
- `kana_char_mask`（query 側と共有しうる純粋関数）は search.rs に残置
- build.rs は `super::{SearchEngine, kana_char_mask}` 経由で親 private を利用し、
  crate 可視性を広げない。move で不要になった 5 import（normalize_entry_key /
  name_char_mask / file_char_mask / lower_file_name / to_lower_folded）を search.rs から除去
- snotra-core/CLAUDE.md の search.rs 索引に build.rs を追記

不変条件: 並列 Vec 同長・kana 系 2 本の {0, entries.len()}・v4 skip / v3 fallback・
`new()` は migemo 有効相当。公開コンストラクタ署名と IndexCache version/wire format は不変。

検証: cargo test -p snotra-core（466 passed, 9 ignored。kana 条件付き構築 3 テスト・
index_cache_on_disk_format_is_stable=golden bytes 不変を含む）/ clippy /
rustdoc(-D warnings) / cargo check -p snotra（下流コンパイル）/ governance:check G1..G9。
