+++ snotra-core/src/search.rs
@@ -15,12 +15,14 @@ use rayon::prelude::*;
 
 use crate::config::{SearchConfig, SearchHistoryNormalizationConfig};
 use crate::history::HistoryStore;
-use crate::indexer::{AppEntry, normalize_entry_key};
+use crate::indexer::AppEntry;
 use crate::query::{
-    char_bitmask, file_char_mask, lower_file_name, name_char_mask, normalize_history_query_key,
-    normalize_query, to_kana, to_lower_folded,
+    char_bitmask, normalize_history_query_key, normalize_query, to_kana,
 };
 use crate::ui_types::SearchResult;
+
+// 構築処理（Wave 1/2・kana マスク・IndexCache 復元・全コンストラクタ）は子モジュールへ分離（#598）。
+mod build;
 
 const GLOBAL_WEIGHT: i64 = 5;
 const QUERY_WEIGHT: i64 = 20;
@@ -219,80 +221,6 @@ struct QueryPlan<'a> {
     path_history_key: Option<String>,
 }
 
-/// Wave 1 の出力: `(lower_names, lower_file_names, normalized_keys, kana_lower_names)`。
-/// いずれも構築後に伸長しないため `Box<str>` で保持する（容量ワード 8B/要素を節約）。
-type Wave1Strings = (Vec<Box<str>>, Vec<Option<Box<str>>>, Vec<Box<str>>, Vec<Box<str>>);
-
-/// Wave 1: entries から文字列正規化データを並列構築する。
-/// lower_names / lower_file_names / normalized_keys / kana_lower_names は
-/// entries への純粋な map であり相互依存がないため rayon::join で並列構築する。
-/// `migemo_enabled` が false の場合、kana_lower_names は空 Vec（migemo 無効ユーザーの
-/// 死蔵メモリを削るため、issue #337）。空 Vec の検索ループ側ガードは search_with_options 参照。
-fn compute_wave1(entries: &[AppEntry], migemo_enabled: bool) -> Wave1Strings {
-    let ((lower_names, lower_file_names), (normalized_keys, kana_lower_names)) = rayon::join(
-        || {
-            rayon::join(
-                || {
-                    entries
-                        .iter()
-                        .map(|e| to_lower_folded(&e.name).into_boxed_str())
-                        .collect::<Vec<_>>()
-                },
-                || {
-                    entries
-                        .iter()
-                        .map(|e| lower_file_name(&e.target_path).map(String::into_boxed_str))
-                        .collect::<Vec<_>>()
-                },
-            )
-        },
-        || {
-            rayon::join(
-                || {
-                    entries
-                        .iter()
-                        .map(|e| normalize_entry_key(&e.target_path).into_boxed_str())
-                        .collect::<Vec<_>>()
-                },
-                || {
-                    // migemo 無効時は kana を構築しない（空 Vec）。
-                    if migemo_enabled {
-                        entries
-                            .iter()
-                            .map(|e| to_kana(&to_lower_folded(&e.name)).into_boxed_str())
-                            .collect::<Vec<_>>()
-                    } else {
-                        Vec::new()
-                    }
-                },
-            )
-        },
-    );
-    (lower_names, lower_file_names, normalized_keys, kana_lower_names)
-}
-
-/// Wave 2: lower_names / lower_file_names からビットマスクを並列構築する。
-/// to_lower_folded が主要な Latin アクセントを ASCII へ折り畳み済み（é→e）のため、
-/// ここで非 ASCII として残る name は典型的には CJK・アラビア文字など。
-/// u64::MAX は任意の query_mask に対して (query_mask & u64::MAX) == query_mask を
-/// 満たすため、これらのエントリはクエリに依らず bitmask pre-filter を常に通過する。
-fn compute_wave2(
-    lower_names: &[Box<str>],
-    lower_file_names: &[Option<Box<str>>],
-) -> (Vec<u64>, Vec<u64>) {
-    rayon::join(
-        || lower_names.iter().map(|n| name_char_mask(n)).collect::<Vec<_>>(),
-        || {
-            // None → 0: entries without a file_name cannot match via the file_name path,
-            // so failing the bitmask check (and being skipped when the name also fails) is correct.
-            lower_file_names
-                .iter()
-                .map(|n| file_char_mask(n.as_deref()))
-                .collect::<Vec<_>>()
-        },
-    )
-}
-
 /// kana の Unicode scalar value を 64 bit に写す損失あり存在マスク。
 /// 同じ文字は必ず同じ bit になるため、mask 不一致なら kana substring は不成立と分かる。
 /// 衝突は候補を余計に通すだけで、kana マッチを棄却しない。
@@ -301,137 +229,7 @@ fn kana_char_mask(kana: &str) -> u64 {
     kana.chars().fold(0, |mask, ch| mask | (1u64 << ((ch as u32) & 63)))
 }
 
-/// migemo 有効時の kana pre-filter 用並列 Vec を構築する。kana 未構築時は空 Vec を保つ。
-fn compute_kana_char_masks(kana_lower_names: &[Box<str>]) -> Vec<u64> {
-    kana_lower_names.iter().map(|name| kana_char_mask(name)).collect()
-}
-
 impl SearchEngine {
-    /// 全並列 Vec の長さが entries と一致することを検証し、Self を組み立てる。
-    fn assemble(
-        entries: Vec<AppEntry>,
-        lower_names: Vec<Box<str>>,
-        lower_file_names: Vec<Option<Box<str>>>,
-        normalized_keys: Vec<Box<str>>,
-        char_masks: Vec<u64>,
-        file_name_char_masks: Vec<u64>,
-        kana: (Vec<Box<str>>, Vec<u64>),
-    ) -> Self {
-        let (kana_lower_names, kana_char_masks) = kana;
-        debug_assert!(
-            lower_names.len() == entries.len()
-                && lower_file_names.len() == entries.len()
-                && normalized_keys.len() == entries.len()
-                && char_masks.len() == entries.len()
-                && file_name_char_masks.len() == entries.len(),
-            "SearchEngine: all parallel Vecs must have the same length as entries"
-        );
-        // kana 系 Vec は {0, entries.len()} を許す（migemo 無効時は空 Vec、issue #337）。
-        debug_assert!(
-            (kana_lower_names.is_empty() && kana_char_masks.is_empty())
-                || (kana_lower_names.len() == entries.len()
-                    && kana_char_masks.len() == entries.len()),
-            "SearchEngine: kana parallel Vecs must both be empty or match entries length"
-        );
-        Self {
-            entries,
-            lower_names,
-            lower_file_names,
-            normalized_keys,
-            char_masks,
-            file_name_char_masks,
-            kana_lower_names,
-            kana_char_masks,
-            prev_query: String::new(),
-            prev_candidates: Vec::new(),
-            prev_mode: None,
-            prev_kana_query: None,
-        }
-    }
-
-    /// kana_lower_names を**常に**構築する（migemo 有効相当）。テスト・ベンチ・convenience 用。
-    /// 本番のインデックス構築は config 由来の migemo フラグを渡す [`Self::new_with_migemo`] /
-    /// [`Self::new_with_cached_masks`] を使い、migemo 無効時は kana を構築しない（issue #337）。
-    pub fn new(entries: Vec<AppEntry>) -> Self {
-        Self::new_with_migemo(entries, true)
-    }
-
-    /// `migemo_enabled` に応じて kana_lower_names の構築要否を決めて構築する。
-    /// false のとき kana は空 Vec（migemo 無効ユーザーの死蔵メモリ ~2.1–2.7MB/50k を削る）。
-    pub fn new_with_migemo(entries: Vec<AppEntry>, migemo_enabled: bool) -> Self {
-        let (lower_names, lower_file_names, normalized_keys, kana_lower_names) =
-            compute_wave1(&entries, migemo_enabled);
-        let (char_masks, file_name_char_masks) =
-            compute_wave2(&lower_names, &lower_file_names);
-        let kana_char_masks = compute_kana_char_masks(&kana_lower_names);
-        Self::assemble(
-            entries,
-            lower_names,
-            lower_file_names,
-            normalized_keys,
-            char_masks,
-            file_name_char_masks,
-            (kana_lower_names, kana_char_masks),
-        )
-    }
-
-    /// キャッシュから読み込んだデータを使って SearchEngine を構築する。
-    ///
-    /// - `char_masks` / `file_name_char_masks`: Wave 2 の再計算をスキップ
-    /// - `cached_lower_names` / `cached_lower_file_names` / `cached_normalized_keys`:
-    ///   v4+ キャッシュヒット時に Some → Wave 1 の再計算もスキップ（A-3）
-    ///   v3 フォールバック時は None → Wave 1 を通常通り並列実行
-    /// - `migemo_enabled`: false のとき kana_lower_names を構築しない（空 Vec、issue #337）。
-    ///   v4 パス（再計算）・v3 フォールバックの**両方**でこのフラグを反映する。
-    pub fn new_with_cached_masks(
-        entries: Vec<AppEntry>,
-        char_masks: Vec<u64>,
-        file_name_char_masks: Vec<u64>,
-        cached_lower_names: Option<Vec<String>>,
-        cached_lower_file_names: Option<Vec<Option<String>>>,
-        cached_normalized_keys: Option<Vec<String>>,
-        migemo_enabled: bool,
-    ) -> Self {
-        let (lower_names, lower_file_names, normalized_keys, kana_lower_names) =
-            if let (Some(ln), Some(lfn), Some(nk)) =
-                (cached_lower_names, cached_lower_file_names, cached_normalized_keys)
-            {
-                // A-3: v4 キャッシュヒット → Wave 1 完全スキップ（kana_lower_names は毎起動再計算）。
-                // キャッシュ由来の Vec<String> を Box<str> へ移す。postcard デシリアライズ後の
-                // String は capacity == len のため into_boxed_str は再アロケーションを伴わない。
-                // migemo 無効時は kana を再計算せず空 Vec のままにする。
-                let kana = if migemo_enabled {
-                    entries
-                        .par_iter()
-                        .map(|e| to_kana(&to_lower_folded(&e.name)).into_boxed_str())
-                        .collect::<Vec<_>>()
-                } else {
-                    Vec::new()
-                };
-                let ln = ln.into_iter().map(String::into_boxed_str).collect::<Vec<_>>();
-                let lfn = lfn
-                    .into_iter()
-                    .map(|o| o.map(String::into_boxed_str))
-                    .collect::<Vec<_>>();
-                let nk = nk.into_iter().map(String::into_boxed_str).collect::<Vec<_>>();
-                (ln, lfn, nk, kana)
-            } else {
-                // v3 フォールバック: Wave 1 を並列実行（migemo フラグを反映）
-                compute_wave1(&entries, migemo_enabled)
-            };
-
-        let kana_char_masks = compute_kana_char_masks(&kana_lower_names);
-        Self::assemble(
-            entries,
-            lower_names,
-            lower_file_names,
-            normalized_keys,
-            char_masks,
-            file_name_char_masks,
-            (kana_lower_names, kana_char_masks),
-        )
-    }
-
     #[inline]
     fn entry_view(&self, i: usize) -> EntryView<'_> {
         EntryView {
